+++ xenon-common/src/test/java/com/vmware/xenon/services/common/TestLuceneDocumentIndexService.java
@@ -446,7 +446,7 @@ public class TestLuceneDocumentIndexService extends BasicReportTestCase {
         Operation get = Operation.createGet(h.getUri()).setCompletion(
                 (o, e) -> {
                     if (e != null) {
-                        host.failIteration(e);
+                        this.host.failIteration(e);
                         return;
                     }
                     ServiceDocumentQueryResult r = o.getBody(ServiceDocumentQueryResult.class);
@@ -457,14 +457,14 @@ public class TestLuceneDocumentIndexService extends BasicReportTestCase {
                         }
                     }
                     if (count != beforeState.size()) {
-                        host.failIteration(new IllegalStateException("Unexpected result:"
+                        this.host.failIteration(new IllegalStateException("Unexpected result:"
                                 + Utils.toJsonHtml(r)));
                     } else {
-                        host.completeIteration();
+                        this.host.completeIteration();
                     }
                 });
         h.queryServiceUris(EnumSet.of(ServiceOption.FACTORY_ITEM), false, get);
-        host.testWait();
+        this.host.testWait();
     }
 
     @Test
@@ -737,7 +737,7 @@ public class TestLuceneDocumentIndexService extends BasicReportTestCase {
             o.setBody(body);
         };
         Consumer<Operation> setBodyMinimal = (o) -> {
-            MinimalTestServiceState body = (MinimalTestServiceState) host.buildMinimalTestState();
+            MinimalTestServiceState body = (MinimalTestServiceState) this.host.buildMinimalTestState();
             o.setBody(body);
         };
 
@@ -807,12 +807,12 @@ public class TestLuceneDocumentIndexService extends BasicReportTestCase {
                 patchExpiration(factoryUri, services, expTime, expectedCount);
                 this.host.log("All example services expired");
 
-                Date exp = host.getTestExpiration();
+                Date exp = this.host.getTestExpiration();
                 while (exp.after(new Date())) {
                     boolean isConverged = true;
                     // confirm services are stopped
                     for (URI u : services.keySet()) {
-                        ProcessingStage s = host.getServiceStage(u.getPath());
+                        ProcessingStage s = this.host.getServiceStage(u.getPath());
                         if (s != null && s != ProcessingStage.STOPPED) {
                             isConverged = false;
                         }
@@ -891,7 +891,7 @@ public class TestLuceneDocumentIndexService extends BasicReportTestCase {
                         maintExpSetBody, factoryUri);
 
                 // do not do anything on the services, rely on the maintenance interval to expire them
-                exp = host.getTestExpiration();
+                exp = this.host.getTestExpiration();
                 while (exp.after(new Date())) {
                     stats = this.host.getServiceState(null, ServiceStats.class, luceneStatsUri);
                     ServiceStat maintExpiredCount = stats.entries
@@ -966,7 +966,7 @@ public class TestLuceneDocumentIndexService extends BasicReportTestCase {
         this.host.testStart(count);
         for (URI u : services.keySet()) {
             this.host.send(Operation.createGet(u).setCompletion((o, e) -> {
-                host.completeIteration();
+                this.host.completeIteration();
             }));
 
             if (!sendDelete) {
@@ -975,7 +975,7 @@ public class TestLuceneDocumentIndexService extends BasicReportTestCase {
             // if expiration is in the past also send a DELETE, to once again make sure its completed
             this.host.send(Operation.createDelete(u).setBody(new ServiceDocument())
                     .setCompletion((o, e) -> {
-                        host.completeIteration();
+                        this.host.completeIteration();
                     }));
         }
         this.host.testWait();
@@ -1032,7 +1032,7 @@ public class TestLuceneDocumentIndexService extends BasicReportTestCase {
         URI factoryUri = UriUtils.buildUri(this.host,
                 ExampleFactoryService.SELF_LINK);
 
-        Map<URI, ExampleServiceState> exampleStates = host.doFactoryChildServiceStart(null,
+        Map<URI, ExampleServiceState> exampleStates = this.host.doFactoryChildServiceStart(null,
                 count,
                 ExampleServiceState.class,
                 (o) -> {
@@ -1124,7 +1124,7 @@ public class TestLuceneDocumentIndexService extends BasicReportTestCase {
 
         URI factoryUri = UriUtils.buildUri(this.host,
                 ExampleFactoryService.SELF_LINK);
-        Map<URI, ExampleServiceState> exampleStates = host.doFactoryChildServiceStart(null,
+        Map<URI, ExampleServiceState> exampleStates = this.host.doFactoryChildServiceStart(null,
                 serviceCount,
                 ExampleServiceState.class,
                 (o) -> {